WorkspaceView in jupyter: back button Fixes #421
